header filter tests are passing
